+++ src/entities/promotions/composables/index.ts
@@ -1,0 +1,2 @@
+export * from "./useLoadTournaments";
+export * from "./useLoadPromo";